support Enterprise Linux
Enterprise Linux distributions such as Red Hat
have a slightly different /proc/net/dev file
format than Debian systems. The difference is
that it is possible for no whitespace to be between
the interface name and the bytes field e.g.
eth0:132342342342 23234222 ...

Debian systems always have the whitespace separator
and pps was written on a Debian system:
eth1: 20607859609382 39372141253

This throws off awk's field numbers. This patch
takes account of this.